refactor(web): extract DesignSystemProjectPanel into its own orchestrator + slice hooks
Moves the ~935-line DesignSystemProjectPanel component out of
FileWorkspace.tsx into components/DesignSystemProjectPanel.tsx, mirroring
the MemorySection orchestrator pattern. Its kit-action cluster (DESIGN.md
load/save, refresh/download kit, publish/default toggles, color/logo/image
edits) moves into features/file-workspace/hooks/useDesignSystemKitActions
behind a new DesignSystemKitActionsPort bound in dependencies.ts, with a
window.confirm bridge added at providers/dom.ts. The per-section review-card
cluster (state + handlers, currently unwired into any render path per the
panel's own "review flow is intentionally gone" comment) moves into
useDesignSystemReviewCards + a DesignSystemReviewCard dumb component,
preserved rather than deleted to stay behavior-identical with the
pre-extraction source.

FileWorkspace.tsx drops from 4029 to 3038 lines.
