+++ apps/web/src/components/FileWorkspace.tsx
@@ -8,11 +8,10 @@ import {
   type ReactNode,
 } from 'react';
 import { Button } from '@open-design/components';
-import type { DesignSystemEditClickProps, TrackingProjectKind } from '@open-design/contracts/analytics';
+import type { TrackingProjectKind } from '@open-design/contracts/analytics';
 import { useAnalytics } from '../analytics/provider';
 import {
   trackFileManagerClick,
-  trackDesignSystemEditClick,
   trackFileUploadResult,
   trackPageView,
   trackTabLauncherClick,
@@ -28,11 +27,8 @@ import {
   fetchProjectFolders,
   applyLibraryAsset,
   createProjectFolder,
-  deleteDesignSystemDraft,
   deleteProjectFolder,
   renameProjectFile,
-  startDesignSystemTokenContractRebuildJob,
-  updateDesignSystemDraft,
   type UploadProjectFilesResult,
   uploadProjectFiles,
   writeProjectBase64File,
@@ -40,25 +36,9 @@ import {
 } from '../providers/registry';
 import { setPendingDesignSystemCreateEntry } from '../analytics/ds-create-entry';
 import { navigate } from '../router';
-import { downloadDesignSystemArchive, downloadProjectArchive } from '../runtime/exports';
-import { finalizeBrandProject } from '../runtime/brands';
-import { deriveFileOps } from '../runtime/file-ops';
-import {
-  deleteBrandImage,
-  deleteBrandLogo,
-  readDesignMd,
-  updateBrandColor,
-} from '../runtime/kit-edit';
-import { latestTodosFromEvents } from '../runtime/todos';
 import { deliverableSlideNavForActiveFile, isSlideNavDeliverableNow } from '../runtime/slide-nav';
-import { useDesignKit, hostnameOf, type KitColor } from '../runtime/design-kit';
-import { useKitModuleUpload } from '../runtime/kit-upload';
-import {
-  DesignKitView,
-  type DesignKitActionFeedbackTone,
-  type DesignKitEditFocusRequest,
-  type HeaderMenuAction,
-} from './DesignKitView';
+import { type DesignKitEditFocusRequest } from './DesignKitView';
+import { DesignSystemProjectPanel } from './DesignSystemProjectPanel';
 import {
   type AgentEvent,
   type AgentInfo,
@@ -95,7 +75,6 @@ import {
   type BrowserPageInfo,
 } from './DesignBrowserPanel';
 import type { PluginFolderAgentAction } from './design-files/pluginFolderActions';
-import { designSystemGithubEvidenceState, repoConnectCopy } from './design-system-github-evidence';
 import { APP_CHROME_FILE_ACTIONS_ID } from './AppChromeHeader';
 import { FileViewer, LiveArtifactViewer } from './FileViewer';
 import { Icon, type IconName } from './Icon';
@@ -104,7 +83,6 @@ import { TabLauncherMenu } from './workspace/TabLauncherMenu';
 import { buildLauncherActions, type LauncherContext } from './workspace/tab-launcher';
 import { SideChatTab, type ActiveConversationChatState } from './workspace/SideChatTab';
 import { TerminalViewer } from './workspace/TerminalViewer';
-import { MissingBrandFontsBanner } from './MissingBrandFontsBanner';
 import { LibraryPicker } from './LibraryPicker';
 import { QuestionsPanel } from './QuestionsPanel';
 import { QuickSwitcher } from './QuickSwitcher';
@@ -125,49 +103,28 @@ import {
   browserTabIndex,
   browserTabsFromState,
   BROWSER_TAB_PREFIX,
-  buildDesignSystemReviewSections,
   colorHexFromBrandJson,
   colorHexFromDesignMd,
-  DesignSystemInlinePreview,
-  DesignSystemProjectLoading,
   DESIGN_FILES_TAB,
   DESIGN_SYSTEM_CARD_MANIFEST_OPTIONAL_STRING_FIELDS,
   DESIGN_SYSTEM_GUIDANCE_FILES,
   DESIGN_SYSTEM_IMAGE_OR_FONT_EXTENSIONS,
   DESIGN_SYSTEM_TAB,
-  designMdBodyWithColor,
   designSystemBasename,
   designSystemFallbackReviewSections,
   designSystemFileOpBelongsToSection,
-  designSystemGenerationProgress,
-  designSystemGenerationReviewHasStarted,
   designSystemGuidanceSort,
   designSystemHasSourceContext,
-  designSystemInitialGenerationSteps,
   designSystemManifestCardError,
   designSystemPathMatchesSection,
   designSystemRelatedFilesForCategory,
-  designSystemReviewAgentTaskLabel,
   designSystemReviewArtifactSort,
   designSystemReviewCategoryRank,
-  designSystemReviewGroups,
-  designSystemReviewNeedsAttention,
-  designSystemReviewPreviewDisplay,
   designSystemReviewSubtitle,
-  designSystemReviewTimeLabel,
   designSystemReviewTitleFromPath,
-  designSystemSectionActivity,
   designSystemSectionActivityLabel,
-  designSystemSectionChangedAfterReview,
-  designSystemSectionEditableFile,
   designSystemSectionPhaseLabel,
-  designSystemSectionPreviewFile,
-  designSystemSectionRunningNotice,
-  designSystemSectionStatus,
-  designSystemSectionStatusClass,
-  designSystemSectionStatusLabel,
   designSystemSectionTodo,
-  designSystemSectionVisibleDuringGeneration,
   designSystemTodoActivityPhase,
   designSystemTodoBelongsToSection,
   designSystemTodoRank,
@@ -175,7 +132,6 @@ import {
   documentTemplateScenarioKey,
   formatWorkspaceSnapshotElapsed,
   inferDesignSystemReviewCategory,
-  initialDesignKitColorHex,
   initialMarkdownDocument,
   isBrowserTabId,
   isDesignSystemAssetFile,
@@ -194,8 +150,6 @@ import {
   lastWorkspaceTabId,
   maxBrowserTabSequence,
   nextMarkdownDocumentPath,
-  normalizeDesignKitHex,
-  normalizeDesignSystemPath,
   normalizeProjectFilePath,
   optionalDesignSystemManifestString,
   orderWorkspaceTabs,
@@ -204,24 +158,14 @@ import {
   reanchorBrowserTabsToCurrentOrder,
   sameFileName,
   scrollWorkspaceTabsWithWheel,
-  slugForTestId,
   Tab,
   tabDropEdgeFromEvent,
   truncateDesignSystemActivityText,
-  useWiredDesignSystemCardManifest,
   type BrowserWorkspaceTab,
-  type DesignSystemGenerationStep,
-  type DesignSystemProjectSection,
-  type DesignSystemProjectSectionReview,
   type DesignSystemReviewAgentTask,
-  type DesignSystemReviewCategory,
   type DesignSystemReviewDecision,
   type DesignSystemReviewDetails,
   type DesignSystemReviewEntry,
-  type DesignSystemReviewPreviewDisplay,
-  type DesignSystemSectionActivity,
-  type DesignSystemSectionActivityPhase,
-  type DesignSystemSectionStatus,
   type TabDropEdge,
   type TranslateFn,
   type WorkspaceOrderedTab,
@@ -3091,939 +3035,4 @@ export function FileWorkspace({
   );
 }
 
-function DesignSystemProjectPanel({
-  projectId,
-  system,
-  brandId,
-  editable,
-  files,
-  streaming,
-  activityEvents,
-  onOpenFile,
-  onUploadAssets,
-  onRefreshFiles,
-  defaultDesignSystemId,
-  onSetDefaultDesignSystem,
-  onDesignSystemsRefresh,
-  onDeleteDesignSystemProject,
-  onNeedsWork,
-  designSystemReview,
-  onReviewDecision,
-  onUseDesignSystem,
-  editFocusRequest,
-  onConnectRepo,
-  githubConnected,
-}: {
-  projectId: string;
-  system: DesignSystemSummary;
-  brandId?: string | null;
-  editable: boolean;
-  files: ProjectFile[];
-  streaming: boolean;
-  activityEvents: AgentEvent[];
-  onOpenFile: (name: string) => void;
-  onUploadAssets: () => void;
-  onRefreshFiles: () => Promise<void> | void;
-  defaultDesignSystemId?: string | null;
-  onSetDefaultDesignSystem?: (id: string | null) => Promise<void> | void;
-  onDesignSystemsRefresh?: () => Promise<void> | void;
-  onDeleteDesignSystemProject?: (id: string) => Promise<boolean> | boolean;
-  onNeedsWork?: (
-    sectionTitle: string,
-    feedback: string,
-    files: string[],
-  ) => DesignSystemReviewAgentTask | void;
-  designSystemReview?: ProjectMetadata['designSystemReview'];
-  onReviewDecision?: (
-    sectionTitle: string,
-    decision: DesignSystemReviewDecision,
-    details?: DesignSystemReviewDetails,
-  ) => void;
-  onUseDesignSystem?: (id: string, title: string) => Promise<void> | void;
-  editFocusRequest?: DesignKitEditFocusRequest | null;
-  onConnectRepo?: () => void;
-  githubConnected?: boolean;
-}) {
-  const t = useT();
-  const analytics = useAnalytics();
-  const [reviewDecisions, setReviewDecisions] = useState<Record<string, DesignSystemReviewDecision>>({});
-  const [expandedSections, setExpandedSections] = useState<Record<string, boolean>>({});
-  const [feedbackSection, setFeedbackSection] = useState<string | null>(null);
-  const [feedbackText, setFeedbackText] = useState('');
-  const [status, setStatus] = useState(system.status ?? 'draft');
-  const [statusBusy, setStatusBusy] = useState(false);
-  const [defaultBusy, setDefaultBusy] = useState(false);
-  useEffect(() => {
-    setStatus(system.status ?? 'draft');
-  }, [system.status]);
-  useEffect(() => {
-    const next: Record<string, DesignSystemReviewDecision> = {};
-    for (const [sectionTitle, entry] of Object.entries(designSystemReview ?? {})) {
-      next[sectionTitle] = entry.decision;
-    }
-    setReviewDecisions(next);
-  }, [designSystemReview]);
-
-  // brand.html-style kit for this design system. brand.json keeps rich assets,
-  // while DESIGN.md is the editable text/token contract rendered on top.
-  const [designMdBody, setDesignMdBody] = useState('');
-  const [savingDesignMd, setSavingDesignMd] = useState(false);
-  const [kitActionBusy, setKitActionBusy] = useState<string | null>(null);
-  // Transient feedback for kit edits (upload / refresh / reset / delete) so an
-  // action that previously fired-and-forgot now reports success or failure.
-  const [kitToast, setKitToast] = useState<{ message: string; tone: DesignKitActionFeedbackTone } | null>(null);
-  const notifyKit = useCallback(
-    (tone: DesignKitActionFeedbackTone, message: string) => setKitToast({ tone, message }),
-    [],
-  );
-  const notifyKitLoading = useCallback(
-    (label: string) => notifyKit('loading', label.endsWith('…') || label.endsWith('...') ? label : `${label}...`),
-    [notifyKit],
-  );
-  const [kitReloadKey, setKitReloadKey] = useState(0);
-  const initialDesignMdRef = useRef<string | null>(null);
-  const initialBrandJsonRef = useRef<string | null>(null);
-  const initialBrandJsonLoadedRef = useRef(false);
-  function emitDesignSystemProjectEditClick(
-    element: DesignSystemEditClickProps['element'],
-    module: DesignSystemEditClickProps['module'],
-  ) {
-    trackDesignSystemEditClick(analytics.track, {
-      page_name: 'design_system_project',
-      area: 'design_system_edit',
-      element,
-      module,
-      edit_surface: 'direct_module',
-      artifact_kind: 'design_system',
-      design_system_id: system.id,
-      project_id: projectId,
-    });
-  }
-
-  const refreshKitDependencies = useCallback(async (options?: { finalizeBrand?: boolean }) => {
-    if (options?.finalizeBrand && brandId) {
-      const outcome = await finalizeBrandProject(brandId, projectId);
-      if (!outcome.ok) throw new Error(outcome.error);
-    }
-    setKitReloadKey((k) => k + 1);
-    await Promise.all([
-      Promise.resolve(onRefreshFiles()),
-      Promise.resolve(onDesignSystemsRefresh?.()),
-    ]);
-  }, [brandId, onDesignSystemsRefresh, onRefreshFiles, projectId]);
-
-  useEffect(() => {
-    let cancelled = false;
-    void Promise.all([
-      readDesignMd(projectId),
-      fetchProjectFileText(projectId, 'brand.json', { cache: 'no-store' }),
-    ]).then(([designMd, brandJson]) => {
-      if (cancelled) return;
-      setDesignMdBody(designMd);
-      if (initialDesignMdRef.current === null) initialDesignMdRef.current = designMd;
-      if (!initialBrandJsonLoadedRef.current) {
-        initialBrandJsonRef.current = brandJson;
-        initialBrandJsonLoadedRef.current = true;
-      }
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [projectId, kitReloadKey]);
-  const kitHost = system.provenance?.sourceUrls?.[0]
-    ? hostnameOf(system.provenance.sourceUrls[0])
-    : undefined;
-  const { uploading: kitUploading, uploadModule: kitUploadModule } = useKitModuleUpload({
-    projectId,
-    title: system.title,
-    onUploaded: (module) => {
-      setKitActionBusy(`upload:${module}`);
-      notifyKit('loading', t('ds.uploading'));
-      void refreshKitDependencies({ finalizeBrand: true })
-        .then(() => notifyKit('success', t('ds.uploadDone')))
-        .catch(() => notifyKit('error', t('ds.actionFailed')))
-        .finally(() => setKitActionBusy(null));
-    },
-    onError: () => {
-      setKitActionBusy(null);
-      notifyKit('error', t('ds.uploadFailed'));
-    },
-  });
-  const { kit } = useDesignKit({
-    designSystemId: system.id,
-    title: system.title,
-    projectId,
-    swatches: system.swatches,
-    body: designMdBody,
-    editable,
-    host: kitHost,
-    reloadKey: kitReloadKey,
-  });
-  async function persistDesignMd(nextBody: string) {
-    const updated = await updateDesignSystemDraft(system.id, { body: nextBody });
-    if (!updated) throw new Error(t('ds.actionFailed'));
-    const file = await writeProjectTextFile(projectId, 'DESIGN.md', nextBody);
-    if (!file) throw new Error(t('ds.actionFailed'));
-    setDesignMdBody(nextBody);
-    await refreshKitDependencies();
-  }
-
-  async function saveDesignMd(nextBody: string) {
-    if (kitActionBusy) throw new Error(t('ds.actionFailed'));
-    setSavingDesignMd(true);
-    setKitActionBusy('design-md-save');
-    notifyKit('loading', t('ds.saving'));
-    try {
-      await persistDesignMd(nextBody);
-      notifyKit('success', t('ds.actionDone'));
-    } catch (err) {
-      notifyKit('error', t('ds.actionFailed'));
-      throw err;
-    } finally {
-      setSavingDesignMd(false);
-      setKitActionBusy(null);
-    }
-  }
-
-  async function refreshKit() {
-    if (kitActionBusy) return;
-    setKitActionBusy('refresh');
-    notifyKitLoading(t('ds.refresh'));
-    try {
-      if (brandId) {
-        await refreshKitDependencies({ finalizeBrand: true });
-      } else {
-        const job = await startDesignSystemTokenContractRebuildJob(system.id, { force: true });
-        if (!job) throw new Error(t('ds.actionFailed'));
-        await refreshKitDependencies();
-      }
-      notifyKit('success', t('ds.actionDone'));
-    } catch {
-      notifyKit('error', t('ds.actionFailed'));
-    } finally {
-      setKitActionBusy(null);
-    }
-  }
-
-  async function downloadKit() {
-    if (kitActionBusy) return;
-    setKitActionBusy('download');
-    notifyKitLoading(t('ds.download'));
-    try {
-      await refreshKitDependencies({ finalizeBrand: true });
-      const ok =
-        await downloadProjectArchive({ projectId, fallbackTitle: system.title }) ||
-        await downloadDesignSystemArchive({ designSystemId: system.id, fallbackTitle: system.title });
-      if (!ok) throw new Error(t('ds.actionFailed'));
-      notifyKit('success', t('ds.actionDone'));
-    } catch {
-      notifyKit('error', t('ds.actionFailed'));
-    } finally {
-      setKitActionBusy(null);
-    }
-  }
-
-  // Delete the whole design system from the project tab's "..." menu: remove the
-  // registered design system (so it leaves the Design Systems list) AND its
-  // backing project, then exit the tab. onDeleteDesignSystemProject is App's
-  // handleDeleteProject, which deletes the project, clears local state and
-  // navigates home — so the panel unmounts on success and there's no busy reset
-  // to do in the happy path.
-  async function deleteDesignSystemProject() {
-    if (kitActionBusy || !onDeleteDesignSystemProject) return;
-    const ok = window.confirm(
-      t('ds.deleteProjectConfirm', { title: system.title }),
-    );
-    if (!ok) return;
-    setKitActionBusy('delete');
-    notifyKitLoading(t('ds.deleteProjectAction', { title: system.title }));
-    try {
-      // Delete the backing project first: this navigates home and unmounts the
-      // panel, so the tab exits cleanly instead of briefly rendering an empty
-      // design-system view. Only on success do we drop the registered design
-      // system (so the Design Systems list keeps no ghost row) and refresh that
-      // list. deleteDesignSystemDraft is a no-op (404 → false) for systems that
-      // aren't user-editable; that's fine.
-      const deleted = await onDeleteDesignSystemProject(projectId);
-      if (!deleted) {
-        notifyKit('error', t('ds.actionFailed'));
-        setKitActionBusy(null);
-        return;
-      }
-      await deleteDesignSystemDraft(system.id);
-      await onDesignSystemsRefresh?.();
-    } catch {
-      notifyKit('error', t('ds.actionFailed'));
-      setKitActionBusy(null);
-    }
-  }
-
-  async function changeKitColor(index: number, hex: string) {
-    if (kitActionBusy) throw new Error(t('ds.actionFailed'));
-    const nextHex = normalizeDesignKitHex(hex);
-    if (!nextHex) throw new Error(t('ds.invalidHexColor'));
-    setKitActionBusy('color');
-    notifyKit('loading', t('ds.saving'));
-    try {
-      const ok = await updateBrandColor(projectId, index, nextHex);
-      if (!ok) {
-        const nextBody = designMdBodyWithColor(designMdBody, kit?.colors ?? [], index, nextHex);
-        await persistDesignMd(nextBody);
-      } else {
-        await refreshKitDependencies({ finalizeBrand: true });
-      }
-      notifyKit('success', t('ds.actionDone'));
-    } catch (err) {
-      notifyKit('error', t('ds.actionFailed'));
-      throw err;
-    } finally {
-      setKitActionBusy(null);
-    }
-  }
-
-  async function resetKitColor(index: number) {
-    const originalHex = initialDesignKitColorHex(index, {
-      brandJson: initialBrandJsonRef.current,
-      designMdBody: initialDesignMdRef.current,
-      swatches: system.swatches,
-      currentColors: kit?.colors ?? [],
-    });
-    if (!originalHex) throw new Error(t('ds.noOriginalColor'));
-    await changeKitColor(index, originalHex);
-  }
-
-  async function removeKitLogo(index: number) {
-    if (kitActionBusy) return;
-    setKitActionBusy(`delete-logo:${index}`);
-    notifyKitLoading(t('ds.deleteLogo'));
-    try {
-      const ok = await deleteBrandLogo(projectId, index);
-      if (!ok) throw new Error(t('ds.actionFailed'));
-      await refreshKitDependencies({ finalizeBrand: true });
-      notifyKit('success', t('ds.actionDone'));
-    } catch {
-      notifyKit('error', t('ds.actionFailed'));
-    } finally {
-      setKitActionBusy(null);
-    }
-  }
-
-  async function removeKitImage(index: number) {
-    if (kitActionBusy) return;
-    setKitActionBusy(`delete-image:${index}`);
-    notifyKitLoading(t('ds.deleteImage', { caption: '' }).trim());
-    try {
-      const ok = await deleteBrandImage(projectId, index);
-      if (!ok) throw new Error(t('ds.actionFailed'));
-      await refreshKitDependencies({ finalizeBrand: true });
-      notifyKit('success', t('ds.actionDone'));
-    } catch {
-      notifyKit('error', t('ds.actionFailed'));
-    } finally {
-      setKitActionBusy(null);
-    }
-  }
-
-  const allFileNames = files.map((file) => file.name);
-  const fileByName = new Map(files.map((file) => [file.name, file]));
-  const manifestFile = files.find((file) => normalizeDesignSystemPath(file.name) === '_ds_manifest.json');
-  const manifestFileName = manifestFile?.name ?? null;
-  const manifestCacheBustKey = manifestFile ? Math.round(manifestFile.mtime) : null;
-  const manifestReadFailedLabel = t('ds.manifestReadFailed');
-  const { cardManifest, cardManifestError } = useWiredDesignSystemCardManifest(
-    projectId,
-    system.id,
-    manifestFileName,
-    manifestCacheBustKey,
-    manifestReadFailedLabel,
-  );
-  const fontFiles = allFileNames.filter((name) =>
-    /\.(otf|ttf|woff|woff2)$/i.test(name) || name.toLowerCase().includes('/fonts/'),
-  );
-  const githubEvidence = designSystemGithubEvidenceState(system, allFileNames);
-  const sections = buildDesignSystemReviewSections(allFileNames, fileByName, cardManifest);
-  const published = status === 'published';
-  const isDefault = published && defaultDesignSystemId === system.id;
-  // Strip a trailing "design system" from the title so the heading
-  // "Review <name> design system" does not read redundantly when a system is
-  // already named e.g. "Acme Design System".
-  const systemDisplayName = system.title.replace(/\s*design system$/i, '').trim() || system.title;
-  const activityFileOps = useMemo(() => deriveFileOps(activityEvents), [activityEvents]);
-  const activityTodos = useMemo(() => latestTodosFromEvents(activityEvents), [activityEvents]);
-  const sectionReviews: DesignSystemProjectSectionReview[] = sections.map((section) => {
-    const previewFile = designSystemSectionPreviewFile(section.files, fileByName);
-    const reviewEntry = designSystemReview?.[section.title];
-    const reviewDecision = reviewDecisions[section.title] ?? reviewEntry?.decision;
-    const sectionActivity = designSystemSectionActivity(section, activityFileOps, activityTodos);
-    const changedAfterFeedback = designSystemSectionChangedAfterReview(
-      section.files,
-      fileByName,
-      reviewEntry,
-    );
-    const sectionStatus = designSystemSectionStatus(
-      section,
-      reviewDecision,
-      changedAfterFeedback,
-      sectionActivity,
-    );
-    return {
-      section,
-      previewFile,
-      previewDisplay: designSystemReviewPreviewDisplay(section, previewFile),
-      reviewEntry,
-      sectionActivity,
-      changedAfterFeedback,
-      sectionStatus,
-      sectionStatusLabel: designSystemSectionStatusLabel(t, section, sectionStatus, sectionActivity),
-      reviewTimeLabel: reviewEntry?.updatedAt
-        ? designSystemReviewTimeLabel(t, reviewEntry.updatedAt)
-        : null,
-    };
-  });
-  const generationReviewHasStarted = published || designSystemGenerationReviewHasStarted(sectionReviews);
-  const visibleSectionReviews = streaming && !published && generationReviewHasStarted
-    ? sectionReviews.filter((item) => designSystemSectionVisibleDuringGeneration(item))
-    : sectionReviews;
-  const groupedSectionReviews = designSystemReviewGroups(visibleSectionReviews);
-  const reviewTocGroups = groupedSectionReviews
-    .map((group) => ({
-      title: group.title,
-      items: group.items.map((item) => ({
-        id: `design-system-section-${slugForTestId(`${group.title}:${item.section.title}`)}`,
-        label: item.section.title,
-        statusClass: designSystemSectionStatusClass(item.sectionStatus),
-        statusLabel: item.sectionStatusLabel,
-      })),
-    }))
-    .filter((group) => group.items.length > 0);
-  const creatingInitialDraft = streaming && !published && !brandId;
-  const generationSteps = designSystemInitialGenerationSteps({
-    files,
-    sectionReviews,
-    system,
-    t,
-  });
-  const generationProgress = designSystemGenerationProgress(generationSteps);
-
-  async function togglePublished(nextPublished: boolean) {
-    if (!editable) return;
-    if (nextPublished && !githubEvidence.ready) return;
-    setStatusBusy(true);
-    notifyKitLoading(publishActionLabel);
-    try {
-      const nextStatus = nextPublished ? 'published' : 'draft';
-      const updated = await updateDesignSystemDraft(system.id, { status: nextStatus });
-      if (!updated) throw new Error(t('ds.actionFailed'));
-      setStatus(updated.status ?? nextStatus);
-      await onDesignSystemsRefresh?.();
-      notifyKit('success', t('ds.actionDone'));
-    } catch {
-      notifyKit('error', t('ds.actionFailed'));
-    } finally {
-      setStatusBusy(false);
-    }
-  }
-
-  async function toggleDefault(nextDefault: boolean) {
-    if (!editable) return;
-    if (!onSetDefaultDesignSystem) return;
-    setDefaultBusy(true);
-    notifyKitLoading(nextDefault ? t('dsManager.makeDefault') : t('dsManager.badgeDefault'));
-    try {
-      await onSetDefaultDesignSystem(nextDefault ? system.id : null);
-      notifyKit('success', t('ds.actionDone'));
-    } catch {
-      notifyKit('error', t('ds.actionFailed'));
-    } finally {
-      setDefaultBusy(false);
-    }
-  }
-
-  function markSectionReview(
-    sectionTitle: string,
-    decision: DesignSystemReviewDecision,
-    details?: DesignSystemReviewDetails,
-  ) {
-    setReviewDecisions((current) => ({ ...current, [sectionTitle]: decision }));
-    onReviewDecision?.(sectionTitle, decision, details);
-    if (decision === 'looks-good' && feedbackSection === sectionTitle) {
-      setFeedbackSection(null);
-      setFeedbackText('');
-    }
-  }
-
-  function toggleSection(sectionTitle: string) {
-    setExpandedSections((current) => ({
-      ...current,
-      [sectionTitle]: !(current[sectionTitle] ?? false),
-    }));
-  }
-
-  function openNeedsWorkFeedback(sectionTitle: string, expansionKey: string) {
-    if (!editable) return;
-    setReviewDecisions((current) => ({ ...current, [sectionTitle]: 'needs-work' }));
-    setExpandedSections((current) => ({ ...current, [expansionKey]: true }));
-    setFeedbackSection(sectionTitle);
-    setFeedbackText('');
-  }
-
-  function submitNeedsWorkFeedback(sectionTitle: string, sectionFiles: string[]) {
-    const feedback = feedbackText.trim();
-    if (!feedback) return;
-    const agentTask = onNeedsWork?.(sectionTitle, feedback, sectionFiles);
-    markSectionReview(sectionTitle, 'needs-work', {
-      feedback,
-      files: sectionFiles,
-      ...(agentTask ? { agentTask } : {}),
-    });
-    setFeedbackSection(null);
-    setFeedbackText('');
-  }
-
-  function renderReviewCard(
-    item: DesignSystemProjectSectionReview,
-    instanceId: string,
-    defaultExpanded: boolean,
-  ) {
-    const {
-      section,
-      previewFile,
-      reviewEntry,
-      sectionActivity,
-      changedAfterFeedback,
-      sectionStatus,
-      sectionStatusLabel,
-    } = item;
-    const needsAttention = designSystemReviewNeedsAttention(item);
-    // A section the user marked "Looks good" is validated, so collapse it by
-    // default to show it is done. Gate that on the current status, not just the
-    // stored decision: when a section is regenerated after approval its status
-    // moves back to needs-attention, and it has to reopen so the "review again"
-    // notice and regenerated preview stay visible. Without the needsAttention guard a stale "looks-good" decision
-    // keeps the regenerated section collapsed and the change is easy to miss.
-    // The user can still re-expand with the chevron (expandedSections[instanceId]),
-    // and an active agent run forces it open.
-    const reviewedGood =
-      !needsAttention && (reviewDecisions[section.title] ?? reviewEntry?.decision) === 'looks-good';
-    const expanded =
-      (expandedSections[instanceId] ?? (defaultExpanded && !reviewedGood)) || sectionActivity.running;
-    const sectionSlug = slugForTestId(instanceId);
-    const sectionAnchorId = `design-system-section-${sectionSlug}`;
-    const editableFile = designSystemSectionEditableFile(section, previewFile, fileByName);
-    return (
-      <section
-        id={sectionAnchorId}
-        key={instanceId}
-        className={[
-          'ds-project-section',
-          'ds-project-review-item',
-          `ds-project-review-item--${item.previewDisplay}`,
-          expanded ? 'is-expanded' : 'is-collapsed',
-        ].join(' ')}
-      >
-        <div className="ds-project-section-head">
-          {/* The trigger is a stretched button covering the whole head, so the
-              entire row toggles. It is a sibling of the review action buttons
-              (not a parent), so there are no nested interactive elements. The
-              title below is display-only (pointer-events: none) and lets clicks
-              fall through to this trigger. */}
-          <button
-            type="button"
-            className="ds-project-section-head-trigger"
-            aria-expanded={expanded}
-            aria-label={t(expanded ? 'ds.reviewCollapseSection' : 'ds.reviewExpandSection', { title: section.title })}
-            onClick={() => toggleSection(instanceId)}
-          />
-          <span className="ds-project-section-title">
-            <Icon name={expanded ? 'chevron-down' : 'chevron-right'} size={13} />
-            <span>
-              <strong>{section.title}</strong>
-              <small>{section.subtitle}</small>
-            </span>
-            {!expanded ? (
-              <span
-                className={[
-                  'ds-project-section-state',
-                  'ds-project-section-dot',
-                  designSystemSectionStatusClass(sectionStatus),
-                ].join(' ')}
-                aria-label={sectionStatusLabel}
-                title={sectionStatusLabel}
-              >
-                {needsAttention ? t('ds.reviewNeedsReview') : t('ds.reviewLooksGood')}
-              </span>
-            ) : null}
-          </span>
-          <div className="ds-project-review-actions" aria-label={t('ds.reviewActionsLabel', { title: section.title })}>
-            <button
-              type="button"
-              className={`ghost success ${reviewDecisions[section.title] === 'looks-good' ? 'active' : ''}`}
-              data-testid={`design-system-review-good-${slugForTestId(section.title)}`}
-              onClick={() => {
-                markSectionReview(section.title, 'looks-good');
-                // Collapse on validate, overriding any manual expand so the
-                // section always tidies away once it is marked good.
-                setExpandedSections((current) => ({ ...current, [instanceId]: false }));
-              }}
-            >
-              <Icon name="check" size={13} />
-              {t('ds.reviewLooksGood')}
-            </button>
-            <button
-              type="button"
-              className={`ghost danger ${reviewDecisions[section.title] === 'needs-work' ? 'active' : ''}`}
-              data-testid={`design-system-review-work-${slugForTestId(section.title)}`}
-              onClick={() => openNeedsWorkFeedback(section.title, instanceId)}
-            >
-              <Icon name="comment" size={13} />
-              {t('ds.reviewNeedsWorkEllipsis')}
-            </button>
-            {editableFile ? (
-              <button
-                type="button"
-                className="ghost compact"
-                data-testid={`design-system-review-edit-${sectionSlug}`}
-                title={t('ds.reviewEditFile', { file: editableFile.name })}
-                onClick={() => onOpenFile(editableFile.name)}
-              >
-                <Icon name="edit" size={13} />
-                {t('common.edit')}
-              </button>
-            ) : null}
-            {feedbackSection === section.title ? (
-              <form
-                className="ds-project-feedback-popover"
-                onSubmit={(event) => {
-                  event.preventDefault();
-                  submitNeedsWorkFeedback(section.title, section.files);
-                }}
-              >
-                <label htmlFor={`ds-feedback-${slugForTestId(section.title)}`}>
-                  {t('ds.reviewFeedbackLabel')}
-                </label>
-                <textarea
-                  id={`ds-feedback-${slugForTestId(section.title)}`}
-                  value={feedbackText}
-                  rows={3}
-                  placeholder={t('ds.reviewFeedbackPlaceholder', { title: section.title })}
-                  onChange={(event) => setFeedbackText(event.target.value)}
-                  autoFocus
-                />
-                <div>
-                  <button
-                    type="button"
-                    className="ghost compact"
-                    onClick={() => {
-                      setFeedbackSection(null);
-                      setFeedbackText('');
-                    }}
-                  >
-                    {t('common.cancel')}
-                  </button>
-                  <button
-                    type="submit"
-                    className="primary compact"
-                    disabled={!feedbackText.trim()}
-                  >
-                    {t('chat.send')}
-                  </button>
-                </div>
-              </form>
-              ) : null}
-          </div>
-        </div>
-        {expanded ? (
-          <div className="ds-project-section-body">
-            {sectionActivity.running ? (
-              <div className="ds-project-review-notice is-running">
-                <Icon name="sparkles" size={14} />
-                <span>{designSystemSectionRunningNotice(t, section, sectionActivity)}</span>
-              </div>
-            ) : changedAfterFeedback || sectionActivity.mutated ? (
-              <div className="ds-project-review-notice">
-                <Icon name="check" size={14} />
-                <span>
-                  {changedAfterFeedback
-                    ? t('ds.reviewChangedAfterFeedback')
-                    : t('ds.reviewChangedDuringRun')}
-                </span>
-              </div>
-            ) : null}
-            {reviewEntry?.decision === 'needs-work' && reviewEntry.feedback ? (
-              <div className="ds-project-last-feedback">
-                <Icon name="comment" size={14} />
-                <span>
-                  <strong>{t('ds.reviewLastFeedback')}</strong>
-                  <small>{reviewEntry.feedback}</small>
-                  {reviewEntry.agentTask ? (
-                    <small>{designSystemReviewAgentTaskLabel(t, reviewEntry.agentTask)}</small>
-                  ) : null}
-                </span>
-              </div>
-            ) : null}
-            {previewFile ? (
-              <div className="ds-project-inline-preview">
-                <DesignSystemInlinePreview projectId={projectId} file={previewFile} />
-              </div>
-            ) : (
-              <div className="ds-project-preview-placeholder">
-                <Icon name="sparkles" size={16} />
-                <span>{t('ds.previewGenerating')}</span>
-              </div>
-            )}
-          </div>
-        ) : null}
-      </section>
-    );
-  }
-
-  if (creatingInitialDraft) {
-    return (
-      <div className="ds-project-panel ds-project-panel--generating">
-        <DesignSystemProjectLoading
-          kicker={t('dsManager.tabDesignSystem')}
-          title={t('ds.creatingProjectTitle')}
-          subtitle={t('ds.creatingProjectSubtitle')}
-          progress={generationProgress}
-          progressLabel={t('ds.generationProgressLabel', { progress: generationProgress })}
-        />
-      </div>
-    );
-  }
-
-  // Scaffolding kept around the brand.html kit: publish / default controls in
-  // the kit header, and the publish card + repo / font / manifest warnings above
-  // the modules. The Looks-good / Needs-work review flow is intentionally gone
-  // here — the kit is the single, on-brand view of the system.
-  // The publish lifecycle button stays a visible primary; everything else
-  // (asset refresh/download/reset and the chat-default toggle) folds into the
-  // header's "More" dropdown so the sticky row reads as one clear action.
-  const repoCopy = repoConnectCopy(t, githubConnected);
-  const publishActionLabel = published ? t('ds.unpublishDesignSystem') : t('ds.publishDesignSystem');
-  const extractionRunning = !editable || streaming;
-  const actionsSlot = (
-    <span
-      className="ds-project-publish-trigger"
-      title={
-        !published && !githubEvidence.ready
-          ? t('ds.publishRepoRequiredTitle')
-          : undefined
-      }
-    >
-      <button
-        type="button"
-        className={published ? 'ghost compact' : 'primary'}
-        data-testid="design-system-publish"
-        aria-label={publishActionLabel}
-        title={publishActionLabel}
-        disabled={!editable || statusBusy || (!published && !githubEvidence.ready)}
-        aria-busy={statusBusy || undefined}
-        onClick={() => void togglePublished(!published)}
-      >
-        <Icon name={statusBusy ? 'spinner' : published ? 'check' : 'arrow-up'} size={14} />
-        {published ? t('ds.published') : t('ds.publish')}
-      </button>
-    </span>
-  );
-
-  const headerMenuActions: HeaderMenuAction[] = [
-    {
-      id: 'refresh',
-      label: t('ds.refresh'),
-      icon: 'refresh',
-      onClick: () => {
-        emitDesignSystemProjectEditClick('kit_refresh', 'kit');
-        void refreshKit();
-      },
-      disabled: !editable || Boolean(kitActionBusy) || statusBusy || defaultBusy,
-      loading: kitActionBusy === 'refresh',
-    },
-    {
-      id: 'download',
-      label: t('dsManager.downloadTitle'),
-      icon: 'download',
-      onClick: () => {
-        emitDesignSystemProjectEditClick('kit_download', 'kit');
-        void downloadKit();
-      },
-      disabled: !editable || Boolean(kitActionBusy) || statusBusy || defaultBusy,
-      loading: kitActionBusy === 'download',
-    },
-    ...(published && onSetDefaultDesignSystem
-      ? [
-          {
-            id: 'default',
-            label: isDefault ? t('dsManager.badgeDefault') : t('dsManager.makeDefault'),
-            icon: (isDefault ? 'check' : 'star') as IconName,
-            onClick: () => void toggleDefault(!isDefault),
-            disabled: !editable || statusBusy || defaultBusy || Boolean(kitActionBusy),
-            loading: defaultBusy,
-            active: isDefault,
-          } satisfies HeaderMenuAction,
-        ]
-      : []),
-    ...(onDeleteDesignSystemProject
-      ? [
-          {
-            id: 'delete',
-            label: t('ds.deleteProjectAction', { title: system.title }),
-            icon: 'trash' as IconName,
-            onClick: () => void deleteDesignSystemProject(),
-            disabled: Boolean(kitActionBusy) || statusBusy || defaultBusy,
-            loading: kitActionBusy === 'delete',
-          } satisfies HeaderMenuAction,
-        ]
-      : []),
-  ];
-
-  const topSlot = (
-    <>
-      <div
-        className={`ds-project-extraction-status ${extractionRunning ? 'is-running' : 'is-complete'}`}
-        role="status"
-        data-testid="design-system-extraction-status"
-      >
-        <Icon name={extractionRunning ? 'sparkles' : 'check'} size={15} />
-        <span>
-          <strong>{extractionRunning ? t('ds.extractionRunningTitle') : t('ds.extractionCompleteTitle')}</strong>
-          <small>
-            {extractionRunning
-              ? t('ds.extractionRunningBody')
-              : t('ds.extractionCompleteBody')}
-          </small>
-        </span>
-      </div>
-
-      <div className="ds-project-publish-card ds-project-publish-card--review">
-        <p>
-          {published
-            ? t('ds.publishCardPublished')
-            : t('ds.publishCardDraft')}
-        </p>
-        {published ? (
-          <div className="ds-project-use-row">
-            <span>
-              <strong>{t('ds.useSystemTitle')}</strong>
-              <small>
-                {t('ds.useSystemBody')}
-              </small>
-            </span>
-            <Button
-              variant="primary"
-              onClick={() => onUseDesignSystem?.(system.id, system.title)}
-              disabled={!onUseDesignSystem}
-            >
-              <Icon name="plus" size={14} />
-              {t('ds.createNewDesign')}
-            </Button>
-          </div>
-        ) : null}
-      </div>
-
-      {!githubEvidence.ready ? (
-        <div className="ds-project-warning-card">
-          <Icon name="github" size={16} />
-          <span>
-            <strong>{repoCopy.bannerTitle}</strong>
-            <small>{repoCopy.bannerBody}</small>
-          </span>
-          {onConnectRepo ? (
-            <Button
-              variant="ghost"
-              className="compact"
-              disabled={githubConnected === undefined}
-              onClick={onConnectRepo}
-            >
-              <Icon name="github" size={13} />
-              {repoCopy.buttonLabel}
-            </Button>
-          ) : githubEvidence.hasSourceManifest ? (
-            <Button variant="ghost" className="compact" onClick={() => onOpenFile('context/source-context.md')}>
-              <Icon name="file" size={13} />
-              {t('ds.openSourceContext')}
-            </Button>
-          ) : null}
-        </div>
-      ) : null}
-
-      {editable && fontFiles.length === 0 ? (
-        <MissingBrandFontsBanner projectId={projectId} onUploadAssets={onUploadAssets} />
-      ) : null}
-
-      {cardManifestError ? (
-        <div
-          className="ds-project-warning-card ds-project-warning-card--error"
-          data-testid="design-system-manifest-error"
-          role="alert"
-        >
-          <Icon name="alert-triangle" size={16} />
-          <span>
-            <strong>{t('ds.manifestNeedsAttention')}</strong>
-            <small>{cardManifestError}</small>
-          </span>
-          {manifestFileName ? (
-            <Button variant="ghost" className="compact" onClick={() => onOpenFile(manifestFileName)}>
-              <Icon name="file" size={13} />
-              {t('ds.openManifest')}
-            </Button>
-          ) : null}
-        </div>
-      ) : null}
-    </>
-  );
-
-  return (
-    <div className="ds-project-panel ds-project-panel--kit" data-testid="design-system-project-tab-panel">
-      {kitToast ? (
-        <Toast
-          message={kitToast.message}
-          tone={kitToast.tone}
-          ttlMs={kitToast.tone === 'loading' ? 60000 : 2600}
-          role={kitToast.tone === 'error' ? 'alert' : 'status'}
-          onDismiss={() => setKitToast(null)}
-        />
-      ) : null}
-      {kit ? (
-        <DesignKitView
-          kit={kit}
-          actionsSlot={actionsSlot}
-          headerMenuActions={headerMenuActions}
-          topSlot={topSlot}
-          stickyHeader
-          designMd={{
-            body: designMdBody,
-            saving: savingDesignMd,
-            canEdit: editable,
-            ...(editable
-              ? {
-                  onSave: saveDesignMd,
-                  onOpenFile: () => onOpenFile('DESIGN.md'),
-                }
-              : {}),
-          }}
-          onUploadModule={editable ? kitUploadModule : undefined}
-          onColorChange={editable ? (index, hex) => changeKitColor(index, hex) : undefined}
-          onColorReset={editable ? (index) => resetKitColor(index) : undefined}
-          onDeleteLogo={editable ? (index) => void removeKitLogo(index) : undefined}
-          onDeleteImage={editable ? (index) => void removeKitImage(index) : undefined}
-          onRefresh={editable ? () => void refreshKit() : undefined}
-          onDownload={editable ? () => void downloadKit() : undefined}
-          onEditClick={emitDesignSystemProjectEditClick}
-          uploading={kitUploading}
-          actionBusy={kitActionBusy}
-          onActionFeedback={notifyKit}
-          editFocusRequest={editFocusRequest}
-          dataTestId="design-system-project-kit"
-        />
-      ) : (
-        <DesignSystemProjectLoading
-          kicker={t('dsManager.tabDesignSystem')}
-          title={systemDisplayName}
-          subtitle={t('ds.workspacePreparing')}
-          progressLabel={t('ds.workspaceLoadingLabel')}
-        />
-      )}
-    </div>
-  );
-}
